Refactored project structure to include a separate 'Interfaces' folder
