+++ WebApplication1/WebApplication1/Interfaces/IService.cs
@@ -1,0 +1,10 @@
+namespace WebApplication1.Interfaces;
+
+public interface IService<T>
+{
+    Task<List<T>> GetAllAsync();
+    Task<T> GetByIdAsync(string id);
+    Task CreateAsync(T entity);
+    Task UpdateAsync(string id, T entity);
+    Task RemoveAsync(string id);
+}
