+++ WebApplication1/WebApplication1/Models/Song.cs
@@ -1,7 +1,7 @@
 using System;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
-using WebApplication1.Services;
+using WebApplication1.Interfaces;
 
 namespace WebApplication1.Models
 {
