+++ WebApplication1/WebApplication1/Services/AlbumService.cs
@@ -1,10 +1,10 @@
 using Microsoft.Extensions.Options;
-using WebApplication1.Controllers;
+using WebApplication1.Interfaces;
 using WebApplication1.Models;
 
 namespace WebApplication1.Services
 {
-    public class AlbumServices : CollectionService<Album>, IService<Album>
+    public class AlbumServices : BaseService<Album>, IService<Album>
     {
         public AlbumServices(IOptions<DBSettings> dbSettings)
             : base(dbSettings, dbSettings.Value.AlbumCollectionName)
