+++ WebApplication1/WebApplication1/Services/SongService.cs
@@ -1,10 +1,10 @@
 using Microsoft.Extensions.Options;
-using WebApplication1.Controllers;
+using WebApplication1.Interfaces;
 using WebApplication1.Models;
 
 namespace WebApplication1.Services
 {
-    public class SongServices : CollectionService<Song>, IService<Song>
+    public class SongServices : BaseService<Song>, IService<Song>
     {
         public SongServices(IOptions<DBSettings> dbSettings)
             : base(dbSettings, dbSettings.Value.SongCollectionName)
